Clean-up of logic, abstraction into submodules
